+++ src/Client/Views/BaseView.fs
@@ -34,7 +34,7 @@ let createNavigationTab (pageLink: Routing.Route) (model:Model) (dispatch:Msg-> 
                 //    str (pageLink |> Routing.Route.toString)
                 //else
                 //    pageLink |> Routing.Route.toIcon
-                span [Class "hideUnder775px"][str (pageLink |> Routing.Route.toString)]
+                span [Class "hideUnder775px"][str pageLink.toStringRdbl]
                 span [Class "hideOver775px"][pageLink |> Routing.Route.toIcon]
             ]
 
@@ -47,7 +47,7 @@ let tabRow (model:Model) dispatch (tabs: seq<ReactElement>)=
         Tabs.Props [
             Style [
                 BackgroundColor model.SiteStyleState.ColorMode.BodyBackground
-                OverflowX OverflowOptions.Hidden
+                CSSProp.Custom ("overflow","visible")
             ]
         ]
     ] [
@@ -79,7 +79,17 @@ open Fable.Core.JsInterop
 
 /// The base react component for all views in the app. contains the navbar and takes body and footer components to create the full view.
 let baseViewComponent (model: Model) (dispatch: Msg -> unit) (bodyChildren: ReactElement list) (footerChildren: ReactElement list) =
-    div [   Style [MinHeight "100vh"; BackgroundColor model.SiteStyleState.ColorMode.BodyBackground; Color model.SiteStyleState.ColorMode.Text;]
+    div [
+        OnClick (fun e ->
+            if model.TermSearchState.ShowSuggestions
+                || model.AddBuildingBlockState.ShowUnitTermSuggestions
+                || model.AddBuildingBlockState.ShowUnit2TermSuggestions
+                || model.AddBuildingBlockState.ShowBuildingBlockTermSuggestions
+            then
+                TopLevelMsg.CloseSuggestions |> TopLevelMsg |> dispatch
+        )
+        Style [MinHeight "100vh"; BackgroundColor model.SiteStyleState.ColorMode.BodyBackground; Color model.SiteStyleState.ColorMode.Text;
+    ]
     ] [
         Navbar.navbarComponent model dispatch
         Container.container [
@@ -87,7 +97,7 @@ let baseViewComponent (model: Model) (dispatch: Msg -> unit) (bodyChildren: Reac
         ] [
             br []
             firstRowTabs model dispatch
-            sndRowTabs model dispatch
+            //sndRowTabs model dispatch
 
             if (not model.ExcelState.HasAnnotationTable) then
                 CustomComponents.AnnotationTableMissingWarning.annotationTableMissingWarningComponent model dispatch
@@ -100,23 +110,15 @@ let baseViewComponent (model: Model) (dispatch: Msg -> unit) (bodyChildren: Reac
 
             br []
 
-            Footer.footer [ Props [ExcelColors.colorControl model.SiteStyleState.ColorMode]] [
-                Content.content [
-                    Content.Modifiers [ Modifier.TextAlignment (Screen.All, TextAlignment.Left)]
-                    Content.Props [ExcelColors.colorControl model.SiteStyleState.ColorMode] 
-                ] [
-                    yield! footerChildren
-                    //Button.button [
-                    //    Button.OnClick (fun e ->
-                    //        let e = Browser.Dom.document.getElementById("BlockNameSearch")
-                    //        let content = e.innerHTML
-                    //        e.innerHTML <- content
-                    //    )
-                    //][
-                    //    str "Test"
-                    //]
+            if footerChildren.IsEmpty |> not then
+                Footer.footer [ Props [ExcelColors.colorControl model.SiteStyleState.ColorMode]] [
+                    Content.content [
+                        Content.Modifiers [ Modifier.TextAlignment (Screen.All, TextAlignment.Left)]
+                        Content.Props [ExcelColors.colorControl model.SiteStyleState.ColorMode] 
+                    ] [
+                        yield! footerChildren
+                    ]
                 ]
-            ]
         ]
 
         div [Style [Position PositionOptions.Fixed; Bottom "0"; Width "100%"; TextAlign TextAlignOptions.Center; Color "grey"]][