+++ src/Client/Views/InfoView.fs
@@ -47,19 +47,45 @@ let introductionElement model dispatch =
         ]
     ]
 
+
+
 let getInContactElement (model:Model) dispatch =
     [
         Label.label [Label.Size Size.IsLarge; Label.Props [Style [Color model.SiteStyleState.ColorMode.Accent; MarginLeft "7%"]]] [str "Get In Contact With Us"]
-        
+
+        Columns.columns [Columns.Props [Style [MaxWidth "80%"; Margin "0 auto"]]][
+            Column.column [][
+                Heading.h5 [Heading.IsSubtitle; Heading.Props [Style [Color model.SiteStyleState.ColorMode.Text]]][
+                    str "Swate is part of the dataPLANT organisation."
+                ]
+                Content.content [][
+                    blockquote [][
+                        str "Services and infrastructures to support "
+                        a [Href "https://twitter.com/search?q=%23FAIRData&src=hashtag_click"][ str "#FAIRData" ]
+                        str " science and good data management practices within the plant basic research community. "
+                        a [Href "https://twitter.com/search?q=%23NFDI&src=hashtag_click"] [ str "#NFDI" ]
+                        str "."
+                    ]
+                ]
+            ]
+            Column.column [Column.Width (Screen.All, Column.IsNarrow); Column.CustomClass "flexToCentered"][
+                a [Href "https://nfdi4plants.de/"; Target "_Blank"; Class "nfdiIcon" ] [
+                    Fulma.Image.image [][
+                        img [Src "https://raw.githubusercontent.com/nfdi4plants/Branding/138420e3b6f9ec9e125c1ca8840874b2be2a1262/logos/DataPLANT_logo_minimal_square_bg_darkblue.svg"]
+                    ]
+                ]
+            ]
+        ]
+
         Columns.columns [Columns.Props [Style [MaxWidth "80%"; Margin "0 auto"]]][
             Column.column [][
                 Heading.h5 [Heading.IsSubtitle; Heading.Props [Style [Color model.SiteStyleState.ColorMode.Text]]][
                     str "Got a good idea or just want to get in touch? Reach out to us! "
-                    div [] [a [Href Shared.URLs.CSBTwitterUrl; Target "_Blank"][str "#Twitter @cs_biology"]]
+                    div [] [a [Href Shared.URLs.NFDITwitterUrl; Target "_Blank"][str "#Twitter @nfdi4plants"]]
                 ]
             ]
-            Column.column [Column.Width (Screen.All, Column.IsOneFifth); Column.CustomClass "flexToCentered"][
-                a [Href Shared.URLs.CSBTwitterUrl; Target "_Blank"][
+            Column.column [Column.Width (Screen.All, Column.IsNarrow); Column.CustomClass "flexToCentered"][
+                a [Href Shared.URLs.NFDITwitterUrl; Target "_Blank"][
                     Fa.i [
                         Fa.Size Fa.Fa4x
                         Fa.Brand.Twitter
@@ -68,34 +94,33 @@ let getInContactElement (model:Model) dispatch =
                 ]
             ]
         ]
-        
-        Columns.columns [Columns.Props [Style [MaxWidth "80%"; Margin "0 auto"]]][
-            Column.column [Column.Width (Screen.All, Column.IsOneFifth); Column.CustomClass "hideUnder775px"][
-                a [Href Shared.URLs.CSBWebsiteUrl; Target "_Blank"][
-                    Fa.i [
-                        Fa.CustomClass "myFaBrand myFaCSB"
-                    ][
-                        str "CSB"
-                    ]
-                ]
-            ]
-            Column.column [][
-                Heading.h5 [Heading.IsSubtitle; Heading.Props [Style [Color model.SiteStyleState.ColorMode.Text]]][
-                    str "If you are interested in our other work, check out our "
-                    b [][str "website"]
-                    str "!"
-                ]
-            ]
-            Column.column [Column.Width (Screen.All, Column.IsOneFifth); Column.CustomClass "flexToCentered hideOver775px"][
-                a [Href Shared.URLs.CSBWebsiteUrl; Target "_Blank"][
-                    Fa.i [
-                        Fa.CustomClass "myFaBrand myFaCSB"
-                    ][
-                        str "CSB"
-                    ]
-                ]
-            ]
-        ]
+
+        //Columns.columns [
+        //    Columns.Props [Style [MaxWidth "80%"; Margin "0 auto"]]
+        //    Columns.CustomClass "reverseCols"
+        //][
+        //    Column.column [
+        //        Column.Width (Screen.All, Column.IsNarrow)
+        //        Column.CustomClass "flexToCentered"
+        //    ][
+        //        a [
+        //            Href Shared.URLs.CSBWebsiteUrl; Target "_Blank"
+        //        ][
+        //            Fa.i [
+        //                Fa.CustomClass "myFaBrand myFaCSB"
+        //            ][
+        //                str "CSB"
+        //            ]
+        //        ]
+        //    ]
+        //    Column.column [][
+        //        Heading.h5 [Heading.IsSubtitle; Heading.Props [Style [Color model.SiteStyleState.ColorMode.Text]]][
+        //            str "If you are interested in our other work, check out our "
+        //            b [][str "website"]
+        //            str "!"
+        //        ]
+        //    ]
+        //]
         
         Columns.columns [Columns.Props [Style [MaxWidth "80%"; Margin "0 auto"]]][
             Column.column [][
@@ -106,7 +131,7 @@ let getInContactElement (model:Model) dispatch =
         
                 ]
             ]
-            Column.column [Column.Width (Screen.All, Column.IsOneFifth); Column.CustomClass "flexToCentered"][
+            Column.column [Column.Width (Screen.All, Column.IsNarrow); Column.CustomClass "flexToCentered"][
                 a [Href "https://github.com/nfdi4plants/Swate/issues"; Target "_Blank"][
                     Fa.i [
                         Fa.Brand.Github